Error Fixed
I couldn't use the proc.ExitCode check to zero, because the app only worked inside the debug mode, not in "normal" mode.

The error was very similar in looks with the error I get with the Mustool Oscilloscope (?...)

Here, I had to check if the StandardOutput.EndOfStream was not empty. It was the only way I found.